page numbering style

+++ converter.py
@@ -13,7 +13,7 @@ def parseLine(line):
     return res
 
 
-def convert(file, output, author, ct, nonums):
+def convert(file, output, author, ct, nonums,numstyle):
     sec = "section"
     if nonums:
         sec = "section*"
@@ -25,6 +25,7 @@ def convert(file, output, author, ct, nonums):
     pdf += "\\usepackage{hyperref}\n"
     pdf += "\\hypersetup{colorlinks=true, linkcolor=blue, filecolor=magenta, urlcolor=cyan, pdfpagemode=FullScreen}\n"
     pdf += "\\urlstyle{same}\n"
+    pdf += "\\pagenumbering{" + numstyle + "}\n"
 
     if author != "":
         pdf += "\\author{"+author+"}\n"
@@ -59,6 +60,8 @@ def convert(file, output, author, ct, nonums):
 
                 aftertext += "\\item " + line[spaces+2:]
 
+            elif line == "\\np" or line == "\\np\n":
+                aftertext += "\\newpage\n"
             else:
                 aftertext += parseLine(line) + " \\\\ \n"
 
@@ -67,26 +70,34 @@ def convert(file, output, author, ct, nonums):
     # writeback
     pdf += "\\title{"+title+"}\n\\maketitle\n"
     if ct:
-        pdf += "\\tableofcontents"
+        pdf += "\\tableofcontents\n"
     pdf += "\\newpage\n"
     pdf += aftertext
     pdf += "\\end{document}"
     with open(output+".tex", "w") as f:
         f.write(pdf)
     subprocess.run(['pdflatex', '-interaction=nonstopmode','-quiet', output+".tex"])
-    os.remove(output+".aux")
-    os.remove(output+".log")
-    os.remove(output+".out")
+    subprocess.run(['pdflatex', '-interaction=nonstopmode','-quiet', output+".tex"])
+    if os.path.exists(output+".aux"):
+        os.remove(output+".aux")
+    if os.path.exists(output+".log"):
+        os.remove(output+".log")
+    if os.path.exists(output+".out"):
+        os.remove(output+".out")
+    if os.path.exists(output+".toc"):
+        os.remove(output+".toc")
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("-c", "--contenttable",
                         action="store_true", help="Enable a content table")
-    parser.add_argument("-a", "--author", type=str, default="")
-    parser.add_argument("-i", "--input", type=str, required=True)
-    parser.add_argument("-o", "--output", type=str, default="output.tex")
-    parser.add_argument("--nonums", action="store_true")
+    parser.add_argument("-a", "--author", type=str, default="", help="Sets author in the pdf file")
+    parser.add_argument("-i", "--input", type=str, required=True, help="The markdown inputfile")
+    parser.add_argument("-o", "--output", type=str, default="output", help="The outputfile (without extension) [default=output]")
+    parser.add_argument("--nonums", action="store_true", help="No numbers for sections")
+    parser.add_argument("-s","--numberingstyle", type=str,default="arabic", help="Style of the page numbering [default=arabic] Options={arabic,roman,Roman,alph,Alph}")
     args = parser.parse_args()
+    assert args.numberingstyle in ["arabic","roman","Roman","alph","Alph"], "Numbering style options doesn't exist"
     convert(args.input, args.output, args.author,
-            args.contenttable, args.nonums)
+            args.contenttable, args.nonums,args.numberingstyle)
